requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
